+++ src/main/java/com/lvhongli/dao/HouseRepository.java
@@ -15,9 +15,9 @@ import org.springframework.data.repository.query.Param;
 /**
  * Created by 瓦力.
  */
-public interface HouseRepository extends JpaRepository<House, Long> ,JpaSpecificationExecutor {
+public interface HouseRepository extends JpaRepository<House, Integer> ,JpaSpecificationExecutor {
 
     @Modifying
     @Query("update House as h  set h.cover=:coverId where h.id=:houseId")
-    int updateCover(@Param("coverId") String coverId, @Param("houseId") Long houseId);
+    int updateCover(@Param("coverId") String coverId, @Param("houseId") Integer houseId);
 }
