+++ src/main/java/com/lvhongli/es/ESRepository.java
@@ -4,7 +4,7 @@ package com.lvhongli.es;
 import org.springframework.data.elasticsearch.repository.ElasticsearchRepository;
 
 
-public interface ESRepository extends ElasticsearchRepository<EsHouseDto,Long> {
+public interface ESRepository extends ElasticsearchRepository<EsHouseDto,Integer> {
 
 
 }
